airpor_reports_ns/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.2
cycler==0.11.0
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
drf-jwt==1.19.2
fonttools==4.33.3
gunicorn==20.1.0
kiwisolver==1.4.3
matplotlib==3.5.2
mysqlclient==2.1.0
numpy==1.23.0
packaging==21.3
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
